Refactor: lift the modules several pages share out of the page that owned them
Eight modules lived inside one page while other pages imported them, which
reads as though previews is built on stacks, or stacks on addons. They are
shared infrastructure that happened to be written where it was first needed.

The five API hooks move to src/hooks, the wizard footer and the addon type icon
to src/components. The stack card is split: previews was importing its rail,
pills, meta grid and age helpers to build a card that only looks the same, so
those become a branded entity-card and the stacks card keeps what is about
stacks. use-org-domains goes the other way — it sits in src/hooks and only
stacks has ever called it.
