+++ frontend/src/components/branded/addon-type-icon.tsx
