+++ frontend/src/components/wizard-footer.tsx
