+++ frontend/src/pages/addons/components/addon-list.tsx
@@ -13,7 +13,7 @@ import { cn } from "@/lib/utils";
 import { formatDistanceToNow } from "date-fns";
 import type { PostgresAddon } from "@/api/addons";
 import { StatusPill } from "./status-pill";
-import { AddonTypeIcon } from "./addon-type-icon";
+import { AddonTypeIcon } from "@/components/branded/addon-type-icon";
 import {
   filterAndSortAddons,
   countByBucket,
