+++ frontend/src/pages/addons/components/postgres-create-page.tsx
@@ -48,8 +48,8 @@ import {
   JsonAreaParseError,
 } from "../lib/payload";
 import StickyActionBar from "@/components/sticky-action-bar";
-import { useObjectStores } from "@/pages/object-stores/hooks/use-object-stores";
-import { usePostgresAddons } from "../hooks/use-postgres-addons";
+import { useObjectStores } from "@/hooks/use-object-stores";
+import { usePostgresAddons } from "@/hooks/use-postgres-addons";
 import { eligibleRestoreSources } from "../lib/restore-sources";
 import { BackupConfigFields } from "./backup-config-fields";
 import { RestoreInitFields } from "./restore-init-fields";
