+++ frontend/src/pages/stacks/components/editor/index.tsx
@@ -9,7 +9,7 @@ import { useStacks } from "@/pages/stacks/contexts/stack-context";
 import { Button } from "@/components/ui/button";
 import { Loader2 } from "lucide-react";
 import { useMemo, useState, useEffect, useCallback, useRef } from "react";
-import { usePostgresAddons } from "@/pages/addons/hooks/use-postgres-addons";
+import { usePostgresAddons } from "@/hooks/use-postgres-addons";
 import type { PostgresAddon } from "@/api/addons";
 import { useStackEditSession, type EditSessionTab } from "@/pages/stacks/hooks/use-stack-edit-session";
 import { statusVariant } from "@/components/branded/status-variant";
@@ -44,7 +44,7 @@ import { useBreadcrumb } from "@/hooks/use-breadcrumb";
 import { getCurrentOrganizationId } from "@/lib/common";
 import { useResourceProjects } from "@/hooks/use-resource-projects";
 import { useCurrentUser } from "@/hooks/use-current-user";
-import { useOrgDomains } from "@/hooks/use-org-domains";
+import { useOrgDomains } from "@/pages/stacks/hooks/use-org-domains";
 import { sortIngresses } from "@/pages/stacks/lib/public-endpoints";
 import { convertFormStackToApiStack, FormStackSchema } from "@/pages/stacks/schemas/form-schema";
 import { useToast } from "@/components/ui/use-toast";
