+++ frontend/src/pages/stacks/components/wizard/templates-browser-panel.tsx
@@ -3,7 +3,7 @@ import { Input } from "@/components/ui/input";
 import { Search, ExternalLink } from "lucide-react";
 import { cn } from "@/lib/utils";
 import type { Template } from "@/pages/stacks/data/templates/types";
-import { WizardFooter } from "./wizard-footer";
+import { WizardFooter } from "@/components/wizard-footer";
 
 interface TemplatesBrowserPanelProps {
   templates: Template[];
